[TASK] Introduce Interfaces for Inventory loading
So that they can be exchanged by DI, for example to load inventories from composer
